Move console to console.ts
